add blog folder
